damage indicators are now shown when a character gets hit
need to make it so melle attacks have a delay in applying their damage like they used to have.

projectiless need to be aimed at time of release instead of time of initiating the attack.

need to add checks to see how close target is to its tile at time of aiming to determine automatic dodge